frontend/requirements.txt
harness-featureflags==1.7.0
Flask==2.3.2
requests==2.32.0
boto3==1.35.0
botocore==1.35.0